+++ src/gui/tools/base.hpp
@@ -207,6 +207,7 @@ public:
         Core,
         Draw,
         Shape,
+        Style,
         User
     };